+++ tests/_bundles/myapp/config.py
@@ -1,7 +1,7 @@
-from flask_unchained import AppBundleConfig
+from flask_unchained import BundleConfig
 
 
-class Config(AppBundleConfig):
+class Config(BundleConfig):
     SECRET_KEY = 'change-me'
     APP_KEY = 'app_key'
     VENDOR_KEY1 = 'app_override'