+++ tests/_bundles/vendor_bundle/config.py
@@ -1,6 +1,6 @@
-from flask_unchained import AppBundleConfig
+from flask_unchained import BundleConfig
 
 
-class Config(AppBundleConfig):
+class Config(BundleConfig):
     VENDOR_KEY1 = 'vendor_key1'
     VENDOR_KEY2 = 'vendor_key2'